SmartShader use for Pallete swapping. PVP opponent cards are hidden. Fixed cannon enemy crash. Fixed field restore algorithm. Fixed startup crash. TODO: pvp needs to be rewritten to throw out syncing and just send all state data when the card select screen is closing

+++ BattleNetwork/netplay/bnMatchMakingPacketProcessor.cpp
@@ -44,7 +44,7 @@ void MatchMaking::PacketProcessor::SetNewRemote(const std::string& socketAddress
 
 void MatchMaking::PacketProcessor::SendPacket(const Poco::Buffer<char>& data)
 {
-  client->sendTo(data.begin(), data.size(), remoteAddr);
+  client->sendTo(data.begin(), (int)data.size(), remoteAddr);
 }
 
 const Poco::Net::SocketAddress& MatchMaking::PacketProcessor::GetRemoteAddr()
